+++ xarray/__init__.py
@@ -18,7 +18,16 @@ from .conventions import SerializationWarning, decode_cf
 from .core.alignment import align, broadcast
 from .core.combine import combine_by_coords, combine_nested
 from .core.common import ALL_DIMS, full_like, ones_like, zeros_like
-from .core.computation import apply_ufunc, corr, cov, dot, polyval, unify_chunks, where
+from .core.computation import (
+    apply_ufunc,
+    call_on_dataset,
+    corr,
+    cov,
+    dot,
+    polyval,
+    unify_chunks,
+    where,
+)
 from .core.concat import concat
 from .core.dataarray import DataArray
 from .core.dataset import Dataset
@@ -46,6 +55,7 @@ __all__ = (
     # Top-level functions
     "align",
     "apply_ufunc",
+    "call_on_dataset",
     "as_variable",
     "broadcast",
     "cftime_range",